requirements/base.txt
pytz==2018.9  # https://github.com/stub42/pytz
python-slugify==2.0.1  # https://github.com/un33k/python-slugify
Pillow==5.4.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis
celery==4.2.1  # pyup: < 5.0  # https://github.com/celery/celery
kombu==4.4.0  # https://github.com/celery/kombu
flower==0.9.3  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==2.0.13  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.1  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Your stuff here...
# ------------------------------------------------------------------------------
pygments>=2.3.1,<3.0
pymdown-extensions>=6.0,<7.0
django-markdown>=0.8.4,<1.0
django-markdownx>=2.0.28,<3.0
django-markdownify>=0.8.0,<1.0
django-simple-history>=2.7.0,<3.0
django-htmlmin>=0.11.0,<1.0
django-sequences>=2.2,<3.0
django-currentuser>=0.3.3,<1.0